\name{MMCox-package}
\alias{MMCox-package}
\alias{MMCox}
\docType{package}
\title{
\packageTitle{MMCox}
}
\description{
\packageDescription{MMCox}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MMCox}
\packageIndices{MMCox}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MMCox}

Maintainer: \packageMaintainer{MMCox}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
